adopting changes in related handlers

+++ SmartFarmer.Core/Movement/IFarmerMoveAtHeightDevice.cs
@@ -7,5 +7,5 @@ namespace SmartFarmer.Movement;
 public interface IFarmerMoveAtHeightDevice
 {
     Task<bool> MoveArmAtHeightAsync(double heightInCm, CancellationToken token);
-    Task<int> MoveArmAtMaxHeightAsync(CancellationToken token);
+    Task<double> MoveArmAtMaxHeightAsync(CancellationToken token);
 }
